functionized alignment angle calculation
